+++ app/seeds/__init__.py
@@ -4,8 +4,7 @@ from .servers import seed_servers, undo_servers
 from .user_servers import seed_user_servers, undo_user_servers
 from .channels import seed_channels, undo_channels
 from .messages import seed_messages, undo_messages
-from .user_messages import seed_user_messages, undo_user_messages, seed_channel_messages, undo_channel_messages
-# from .channel_messages import seed_channel_messages, undo_channel_messages
+from .user_channel_messages import seed_user_messages, undo_user_messages, seed_channel_messages, undo_channel_messages
 
 # Creates a seed group to hold our commands
 # So we can type `flask seed --help`
